P2P-187, P2P-289, P2P-210

+++ app/src/main/java/com/p2p/wallet/main/model/TokenConverter.kt
@@ -30,7 +30,7 @@ object TokenConverter {
             price = account.getFormattedPrice(exchangeRate, decimals),
             total = total,
             decimals = decimals,
-            walletBinds = bid,
+            walletBinds = bid.toBigDecimal(),
             color = wallet.color,
             exchangeRate = exchangeRate,
             isHidden = total.compareTo(BigDecimal.ZERO) == 0
